fix callback crash
Creating a FunPtr was the right idea, but peekFunPtr turns it back into
a regular Haskell function, so calling it from C is just as problematic
as before. Instead, use castFunPtrToPtr to obtain the corresponding raw
pointer.

Note that neither this version nor the buggy version is calling
freeHaskellFunPtr anywhere, so this leaks memory.